Moved the test cases of ERC20EXPBase on abstracts dir
