+++ test/index.test.ts
@@ -1,4 +1,3 @@
-import * as ERC20EXPBase from "./ERC20EXPBase/index.test";
 import * as LightWeightERC20EXPBase from "./LightWeightERC20EXPBase/index.test";
 import * as ERC20EXPWhitelist from "./ERC20EXPWhitelist/index.test";
 import * as LightWeightERC20EXPWhitelist from "./LightWeightERC20EXPWhitelist/index.test";
@@ -6,7 +5,6 @@ import * as abstracts from "./abstracts/index.test";
 import * as libraries from "./libraries/index.test";
 
 describe("Scenario", async function () {
-  // ERC20EXPBase.run();
   // LightWeightERC20EXPBase.run();
 
   // ERC20EXPWhitelist.run();
